+++ packages/client/src/ContractFactory.ts
@@ -4,25 +4,24 @@ import { Provider } from '@ethersproject/providers'
 import { Signer } from '@ethersproject/abstract-signer'
 import { ObservableContract, createDecoratedContract } from './utils/contract'
 import { SynchronizedGraphQLClient } from './utils/SynchronizedGraphQLClient'
-import { EthereumConfig } from './Ethereum'
-import { ConfigInjectionToken } from './Config'
+import { ConfigInjectionToken, StrictStreamrClientConfig } from './Config'
 import { EthereumAddress } from '@streamr/utils'
 import { LoggerFactory } from './utils/LoggerFactory'
 
 @scoped(Lifecycle.ContainerScoped)
 export class ContractFactory {
     private readonly graphQLClient: SynchronizedGraphQLClient
-    private readonly ethereumConfig: EthereumConfig
     private readonly loggerFactory: LoggerFactory
+    private readonly config: Pick<StrictStreamrClientConfig, 'contracts'>
 
     constructor(
         graphQLClient: SynchronizedGraphQLClient,
-        @inject(ConfigInjectionToken.Ethereum) ethereumConfig: EthereumConfig,
-        @inject(LoggerFactory) loggerFactory: LoggerFactory
+        @inject(LoggerFactory) loggerFactory: LoggerFactory,
+        @inject(ConfigInjectionToken) config: Pick<StrictStreamrClientConfig, 'contracts'>
     ) {
         this.graphQLClient = graphQLClient
-        this.ethereumConfig = ethereumConfig
         this.loggerFactory = loggerFactory
+        this.config = config
     }
 
     createReadContract<T extends Contract>(
@@ -35,7 +34,7 @@ export class ContractFactory {
             new Contract(address, contractInterface, provider),
             name,
             this.loggerFactory,
-            this.ethereumConfig.maxConcurrentCalls
+            this.config.contracts.maxConcurrentCalls
         )
     }
 
@@ -50,7 +49,7 @@ export class ContractFactory {
             name,
             this.loggerFactory,
             // The current maxConcurrentCalls value is just a placeholder as we don't support concurrent writes (as we don't use nonces).
-            // When we add the support, we should use this.ethereumConfig.maxConcurrentCalls here.
+            // When we add the support, we should use the maxConcurrentCalls option from client config here.
             // Also note that if we'd use a limit of 1, it wouldn't make the concurrent transactions to a sequence of transactions,
             // because the concurrency limit covers only submits, not tx.wait() calls.
             999999