+++ packages/client/src/registry/StreamRegistry.ts
@@ -4,9 +4,9 @@ import StreamRegistryArtifact from '../ethereumArtifacts/StreamRegistryV3Abi.jso
 import { BigNumber } from '@ethersproject/bignumber'
 import { Provider } from '@ethersproject/providers'
 import { scoped, Lifecycle, inject, delay } from 'tsyringe'
-import { EthereumConfig, getAllStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
+import { getAllStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
 import { until } from '../utils/promises'
-import { ConfigInjectionToken, TimeoutsConfig } from '../Config'
+import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
 import { Stream, StreamMetadata } from '../Stream'
 import { NotFoundError } from '../HttpUtil'
 import { StreamID, StreamIDUtils } from '@streamr/protocol'
@@ -79,14 +79,13 @@ export class StreamRegistry {
         @inject(SynchronizedGraphQLClient) private graphQLClient: SynchronizedGraphQLClient,
         @inject(delay(() => StreamRegistryCached)) private streamRegistryCached: StreamRegistryCached,
         @inject(AuthenticationInjectionToken) private authentication: Authentication,
-        @inject(ConfigInjectionToken.Ethereum) private ethereumConfig: EthereumConfig,
-        @inject(ConfigInjectionToken.Timeouts) private timeoutsConfig: TimeoutsConfig
+        @inject(ConfigInjectionToken) private config: Pick<StrictStreamrClientConfig, 'contracts' | '_timeouts'>
     ) {
         this.logger = loggerFactory.createLogger(module)
-        const chainProviders = getAllStreamRegistryChainProviders(ethereumConfig)
+        const chainProviders = getAllStreamRegistryChainProviders(config)
         this.streamRegistryContractsReadonly = chainProviders.map((provider: Provider) => {
             return this.contractFactory.createReadContract<StreamRegistryContract>(
-                toEthereumAddress(this.ethereumConfig.streamRegistryChainAddress),
+                toEthereumAddress(this.config.contracts.streamRegistryChainAddress),
                 StreamRegistryArtifact,
                 provider,
                 'streamRegistry'
@@ -103,7 +102,7 @@ export class StreamRegistry {
         if (!this.streamRegistryContract) {
             const chainSigner = await this.authentication.getStreamRegistryChainSigner()
             this.streamRegistryContract = this.contractFactory.createWriteContract<StreamRegistryContract>(
-                toEthereumAddress(this.ethereumConfig.streamRegistryChainAddress),
+                toEthereumAddress(this.config.contracts.streamRegistryChainAddress),
                 StreamRegistryArtifact,
                 chainSigner,
                 'streamRegistry'
@@ -112,7 +111,7 @@ export class StreamRegistry {
     }
 
     async createStream(streamId: StreamID, metadata: StreamMetadata): Promise<Stream> {
-        const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
+        const ethersOverrides = getStreamRegistryOverrides(this.config)
 
         const domainAndPath = StreamIDUtils.getDomainAndPath(streamId)
         if (domainAndPath === undefined) {
@@ -133,8 +132,10 @@ export class StreamRegistry {
             try {
                 await until(
                     async () => this.streamExistsOnChain(streamId),
-                    this.timeoutsConfig.jsonRpc.timeout,
-                    this.timeoutsConfig.jsonRpc.retryInterval
+                    // eslint-disable-next-line no-underscore-dangle
+                    this.config._timeouts.jsonRpc.timeout,
+                    // eslint-disable-next-line no-underscore-dangle
+                    this.config._timeouts.jsonRpc.retryInterval
                 )
             } catch (e) {
                 throw new Error(`unable to create stream "${streamId}"`)
@@ -157,7 +158,7 @@ export class StreamRegistry {
     // Most likely the contract doesn't make any merging (like we do in Stream#update)?
     async updateStream(streamId: StreamID, metadata: Partial<StreamMetadata>): Promise<Stream> {
         await this.connectToContract()
-        const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
+        const ethersOverrides = getStreamRegistryOverrides(this.config)
         await waitForTx(this.streamRegistryContract!.updateStreamMetadata(
             streamId,
             JSON.stringify(metadata),
@@ -169,7 +170,7 @@ export class StreamRegistry {
     async deleteStream(streamIdOrPath: string): Promise<void> {
         const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)
         await this.connectToContract()
-        const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
+        const ethersOverrides = getStreamRegistryOverrides(this.config)
         await waitForTx(this.streamRegistryContract!.deleteStream(
             streamId,
             ethersOverrides
@@ -339,16 +340,16 @@ export class StreamRegistry {
     async grantPermissions(streamIdOrPath: string, ...assignments: PermissionAssignment[]): Promise<void> {
         return this.updatePermissions(streamIdOrPath, (streamId: StreamID, user: EthereumAddress | undefined, solidityType: BigNumber) => {
             return (user === undefined)
-                ? this.streamRegistryContract!.grantPublicPermission(streamId, solidityType, getStreamRegistryOverrides(this.ethereumConfig))
-                : this.streamRegistryContract!.grantPermission(streamId, user, solidityType, getStreamRegistryOverrides(this.ethereumConfig))
+                ? this.streamRegistryContract!.grantPublicPermission(streamId, solidityType, getStreamRegistryOverrides(this.config))
+                : this.streamRegistryContract!.grantPermission(streamId, user, solidityType, getStreamRegistryOverrides(this.config))
         }, ...assignments)
     }
 
     async revokePermissions(streamIdOrPath: string, ...assignments: PermissionAssignment[]): Promise<void> {
         return this.updatePermissions(streamIdOrPath, (streamId: StreamID, user: EthereumAddress | undefined, solidityType: BigNumber) => {
             return (user === undefined)
-                ? this.streamRegistryContract!.revokePublicPermission(streamId, solidityType, getStreamRegistryOverrides(this.ethereumConfig))
-                : this.streamRegistryContract!.revokePermission(streamId, user, solidityType, getStreamRegistryOverrides(this.ethereumConfig))
+                ? this.streamRegistryContract!.revokePublicPermission(streamId, solidityType, getStreamRegistryOverrides(this.config))
+                : this.streamRegistryContract!.revokePermission(streamId, user, solidityType, getStreamRegistryOverrides(this.config))
         }, ...assignments)
     }
 
@@ -391,7 +392,7 @@ export class StreamRegistry {
             }))
         }
         await this.connectToContract()
-        const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
+        const ethersOverrides = getStreamRegistryOverrides(this.config)
         const txToSubmit = this.streamRegistryContract!.setPermissionsMultipleStreans(
             streamIds,
             targets,